requirements.txt
Django==4.2.27
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-environ==0.12.0
django-cors-headers==4.9.0
psycopg[binary]==3.2.13
gunicorn==22.0.0
whitenoise==6.11.0
django-storages==1.14.6
boto3==1.34.131

# Real-time & Performance
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
celery==5.3.6
django-redis==5.4.0
daphne==4.1.0

# Monitoring & Optimization
django-debug-toolbar==4.2.0
django-silk==5.1.0
sentry-sdk==1.40.0

# Additional utilities
django-extensions==3.2.3
django-filter==24.1
